Make ci/bin/format.dart more idiomatic and move to tools/formatter
