+++ ci/analyze.sh
@@ -70,12 +70,6 @@ analyze \
   --options "$FLUTTER_DIR/analysis_options.yaml" \
   "$FLUTTER_DIR/lib/spirv"
 
-echo "Analyzing ci/"
-analyze \
-  --packages="$FLUTTER_DIR/ci/.dart_tool/package_config.json" \
-  --options "$FLUTTER_DIR/analysis_options.yaml" \
-  "$FLUTTER_DIR/ci"
-
 echo "Analyzing flutter_frontend_server..."
 analyze \
   --packages="$FLUTTER_DIR/flutter_frontend_server/.dart_tool/package_config.json" \
@@ -137,6 +131,12 @@ analyze \
   --packages="$FLUTTER_DIR/tools/clang_tidy/.dart_tool/package_config.json" \
   --options "$FLUTTER_DIR/analysis_options.yaml" \
   "$FLUTTER_DIR/tools/clang_tidy"
+
+echo "Analyzing tools/formatter"
+analyze \
+  --packages="$FLUTTER_DIR/tools/formatter/.dart_tool/package_config.json" \
+  --options "$FLUTTER_DIR/analysis_options.yaml" \
+  "$FLUTTER_DIR/tools/formatter"
 
 # Check that dart libraries conform.
 echo "Checking web_ui api conformance..."
