+++ tools/pub_get_offline.py
@@ -15,7 +15,6 @@ import subprocess
 import sys
 
 ALL_PACKAGES = [
-  os.path.join("src", "flutter", "ci"),
   os.path.join("src", "flutter", "flutter_frontend_server"),
   os.path.join("src", "flutter", "testing", "benchmark"),
   os.path.join("src", "flutter", "testing", "litetest"),
@@ -24,6 +23,7 @@ ALL_PACKAGES = [
   os.path.join("src", "flutter", "tools", "android_lint"),
   os.path.join("src", "flutter", "tools", "clang_tidy"),
   os.path.join("src", "flutter", "tools", "const_finder"),
+  os.path.join("src", "flutter", "tools", "formatter"),
   os.path.join("src", "flutter", "tools", "githooks"),
   os.path.join("src", "flutter", "tools", "licenses"),
 ]
